+++ src/components/Paragraph.ts
@@ -5,14 +5,14 @@
 import { Child, DE } from "@donutteam/document-builder";
 
 //
-// Exports
+// Component
 //
 
-export function Paragraph(child : Child) : DE
+export function Paragraph(child: Child)
 {
 	if (child == null)
 	{
-		return new DE(null, null);
+		return null;
 	}
 
 	return new DE("p", "component-paragraph", child);