+++ src/components/head/Preload.ts
@@ -3,12 +3,15 @@
 //
 
 import { DE } from "@donutteam/document-builder";
+import { z } from "zod";
 
 //
 // Component
 //
 
-export type PreloadType = "audio" | "document" | "embed" | "fetch" | "font" | "image" | "object" | "script" | "style" | "track" | "worker" | "video";
+export const PreloadTypeSchema = z.enum([ "audio", "document", "embed", "fetch", "font", "image", "object", "script", "style", "track", "worker", "video" ]);
+
+export type PreloadType = z.infer<typeof PreloadTypeSchema>;
 
 export function Preload(type: PreloadType, getHref: () => string)
 {